requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pinecone-client
openai